# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RAG API over Diogenes Laertius, Lives of Eminent Philosophers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sections:
    get:
      operationId: listSections
      tags: [corpus]
      summary: List sections
      description: Browse corpus sections, filterable by book and philosopher
      parameters:
        - name: book
          in: query
          required: false
          schema:
            type: integer
        - name: philosopher
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"
  /sections/{id}:
    get:
      operationId: getSection
      tags: [corpus]
      summary: Get a section
      description: A single section by its D.L. reference id (e.g. 1.1.24)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
components:
  schemas:
...
    ErrorMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ExternalLinks:
      type: object
      description: Links to external authority records for a philosopher
      properties:
        qid:
          type: string
          description: Wikidata entity id
        enwiki:
          type: string
          description: English Wikipedia title
...
    Section:
      type: object
      required: [id, urn, book, chapter, section, philosopher, school, text]
      properties:
        id:
          type: string
        urn:
          type: string
        book:
          type: integer
        chapter:
          type: string
        section:
          type: string
        philosopher:
          type: string
        school:
          type: string
        schoolGrc:
          type: string
          description: Curated Greek display form of `school` (greek-names curation); absent when the label has no curated Greek form
        text:
          type: string
          description: Greek text
        textEn:
          type: ["string", "null"]
          description: Aligned English translation (Hicks), when available
        externalLinks:
          $ref: "#/components/schemas/ExternalLinks"
...